Split klog tests into feature-focused modules
